added profile toggle
