+++ client/src/components/App/App.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import ProtectedRoute from './AuthFolder/ProtectedRoute.jsx'
+import ProtectedRoute from "./AuthFolder/ProtectedRoute.jsx";
 import LoginLandingPage from "./LandingPage/LoginLandingPage";
 import { LandingPageProvider } from "./LandingPage/LandingPageContext";
 import StudentsPage from "./StudentsPage/StudentsPage";
