+++ client/src/components/App/StudentsPage/Profile/ProfileOpen.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+const ProfileOpen = () => {
+    return (
+        <>
+        <div>test</div>
+        </>
+    )
+};
+export default ProfileOpen;
