--- client/src/components/App/StudentsPage/ProfileMenu/profileOpen.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-const profileOpen = () => {
-    return (
-        <>
-        </>
-    )
-};
-export default profileOpen;
